# backend/internal/apidocs/openapi.yaml
openapi: 3.0.3
info:
  title: Clothing Store API
  version: 1.0.0
  description: REST API for authentication, catalog, carts, and orders.
servers:
  - url: /
    description: Current server
...
paths:
...
  /api/orders/{id}/cancel:
    put:
      tags: [Orders]
      summary: Cancel the authenticated customer's pending order
      description: Atomically restores stock for every ordered variant and marks the order cancelled. Only a pending order owned by the authenticated customer can be cancelled.
      operationId: cancelOrder
      security: [{ bearerAuth: [] }]
      parameters:
        - $ref: '#/components/parameters/Id'
      responses:
        '200':
          description: Order cancelled and stock restored
          content:
            application/json:
              schema: { $ref: '#/components/schemas/OrderCancellationResponse' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '500': { $ref: '#/components/responses/InternalError' }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT returned by POST /api/auth/login.
  parameters:
    Id:
      name: id
      in: path
      required: true
      schema: { $ref: '#/components/schemas/ObjectId' }
...
  responses:
    BadRequest:
      description: Invalid request
      content:
        text/plain:
          schema: { type: string, example: invalid request body }
    Unauthorized:
      description: Missing or invalid bearer token
      content:
        text/plain:
          schema: { type: string, example: Unauthorized }
    Forbidden:
      description: Admin role required
      content:
        text/plain:
          schema: { type: string, example: Forbidden }
    NotFound:
      description: Resource not found
      content:
        text/plain:
          schema: { type: string, example: Not Found }
...
    InternalError:
      description: Unexpected server error
      content:
        text/plain:
          schema: { type: string, example: Internal Server Error }
  schemas:
    ObjectId:
      type: string
      pattern: '^[a-fA-F0-9]{24}$'
      example: 507f1f77bcf86cd799439011
    Timestamp:
      type: string
      format: date-time
...
    OrderItem:
      type: object
      required: [productId, variantId, productName, color, size, price, quantity, subtotal]
      properties:
        productId: { $ref: '#/components/schemas/ObjectId' }
        variantId: { $ref: '#/components/schemas/ObjectId' }
        productName: { type: string }
        color: { type: string }
        size: { type: string }
        price: { type: number, format: double }
        quantity: { type: integer, minimum: 1 }
        subtotal: { type: number, format: double }
    Order:
      type: object
      required: [id, userId, items, totalPrice, status, createdAt, updatedAt]
      properties:
        id: { $ref: '#/components/schemas/ObjectId' }
        userId: { $ref: '#/components/schemas/ObjectId' }
        items:
          type: array
          items: { $ref: '#/components/schemas/OrderItem' }
        totalPrice: { type: number, format: double }
        status: { type: string, enum: [pending, processing, shipped, delivered, cancelled] }
        createdAt: { $ref: '#/components/schemas/Timestamp' }
        updatedAt: { $ref: '#/components/schemas/Timestamp' }
...
    OrderCancellationResponse:
      type: object
      required: [message, order]
      properties:
        message: { type: string, example: order cancelled successfully }
        order: { $ref: '#/components/schemas/Order' }
...